python/logging/requirements.txt
structlog
loguru
colorlog
pyyaml
fastapi
uvicorn
python-json-logger
